+++ test/support/active_record.rb
@@ -74,6 +74,8 @@ class User < ActiveRecord::Base
   encrypts :city, padding: true
   encrypts :ssn, encode: false
 
+  encrypts :state
+
   has_rich_text :content if respond_to?(:has_rich_text)
 end